Fix for issue #453.
Removed the observation init from MH. This means that the sampler might
start from a state that violates the conditions, but the models is still
in a valid state (ie, valid according to its generative model).

This fix also uncovered another MH bug that was found from the decision
tests. In the MultiDecisionTest example, it was recording decisions
values and expected utilities for non-valid combinations. Meaning, the
expected utility of a decision was a value that was not possible based
on the model.

It turns out that this bug has to do with the update function in MH.
This function was repeatedly visiting elements to update, but not in the
correct order. I couldn't figure out the exact nature of the bug, but
when you have a model with A->B and (A,B)->C, you would get B updated,
then C, then B again. This was incorrect.

The solution is I just rewrote the updateMany function to do a recursive
pass through the dependencies. Visiting elements in generative order is
a repeated operation throughout Figaro (MH, IS, Forward sampling) and we
really should create a unified visiting method for all algorithms.