Ising model example using Gibbs/SFI
Demonstrates potential performance benefits of a Gibbs/VE chooser over
regular Gibbs sampling. In the experimental package for now because
using VE to compute the true probabilities is sometimes unreliable; once
in a while it chooses a poor elimination order and runs out of memory.
Currently addressing this with a built-in kill switch for when VE takes
too long.