Specialized ProbEvidenceSampler for marginal MAP
This sampler doesn't clear all temporary elements, but just the newly
created temporary elements. This prevents the MHCache from
unintentionally being cleared.

Also removed the ability to specify an arbitrary probability of evidence
function because it would have too many restrictions for the user to
care about (i.e. that the universe state must not change, and the
elements in the MH cache must not be cleared).