Normal proposals for univariate continuous elements
In experimental package for now because this needs some thorough testing
for correctness, and I don't have time for that just yet.

Some notes for future work:
It might be appropriate to change some of the defs to lazy vals.
In the long term, we should be able to merge these changes into the
existing library.atomic.continuous package, since these elements have
just been copied over for now.
I haven't added normal proposals for Beta because I figure they probably
wont perform well in the case where a and b are close to 0, since this
distribution is bimodal.
Some of the constructs I created in NormalProposer hint at
inconsistencies between how densities are handled with respect to
element randomness or value. We might want to address this in the
future.