Improvements for normal proposals
Before, we were creating an Apache Commons NormalDistribution each time
we wanted to compute the proposal probability. This ended up being
painfully slow because of overhead associated with instantiating a
random generator for this purpose. We really just want the ability to
compute density and cumulative probability for an arbitrary normal
distribution. Utilities for this were added in the Normal object.

Also added more tests.