RefiningStrategy trait
The most important part of this commit is the ability (and necessity) of
a refining strategy to mark problem components as fully expanded. This
flag makes these candidates ineligible for further refinement.

Also note that any refinement that changes problem structure in a way
that makes the previous solution inapplicable must mark these problems
as unsolved.