Reworked DecompositionStrategy
Decomposition strategies no longer solve problems; solving will be done
separately. This combines some of the funcitonality of
DecompositionStrategy, BackwardChain, and StructuredStrategy.
Ultimately, the goal is to replace all of these. This still needs a
couple more comments, and one or two implementing decomposition
strategies should be added.

Some small changes to Problem and ProblemComponent were made to account
for these changes. For example, it will be the responsibility of solving
strategies to raise subproblem solutions.