Remove ProblemStrategy
The primary reason for this is that for a number of refining strategies,
it doesn't make sense to specify a problem for refinement. In reality,
it may refine components in any number of related problems. What's
important is that the components belong to a single collection.

This involved some refactoring of refining strategies as well. The big
change here was to make BottomUpStrategy take initialComponents as an
explicit list.