Refining strategy rework
Renamed BottomUpStrategy to ExpansionStrategy to avoid confusion. ExpansionStrategy is lazy, taking a depth parameter that defaults to Int.MaxValue. Also reworked TopDownStrategy into FlatStrategy that works similarly, but does not explicitly perform the top-down search (this has been replaced by the FlatStrategy.topDown method).

However, the primary purpose of this change is to correctly implement backtracking for range refinement. Specifically, it is possible that lazy expansion necessitates ranging a component multiple times, and at different depths. For more details, compare to the implementation of LazyValues for LFI (or see the explanation in the LFI paper).

There are still some TODOs here, particularly on the testing side. Some of the previous tests are no longer applicable; these should be revisited and replaced. We should also check cyclic dependencies in the problem graph are correctly resolved. In making these changes, I also broke up the shared inheritance between ExpansionStrategy and FlatStrategy. There is still room for shared code, but this needs to be reevaluated.