Better expansion of recursive subproblems
Recursion within a subproblem prevents some memoization. However, we can still do better than the exponential growth in the problem graph that may result from using no memoization at all. If we greedily take the shallowest expansion of a subproblem that does not result in a cycle, then the number of copies of a subproblem will be at most linear in the depth of expansion.

There are still some improvements to be made here. For example, one could implement a better incremental cycle detection algorithm. We might also split ComponentCollection into lazy and non-lazy versions so that non-lazy algorithms avoid the lazy overhead. The changes made in this commit also need testing.