Fixes for ExpansionStrategy
This primarily addresses the case where refining a subproblem of an expandable component results in a change in the range of the parent, which necessitates expanding subproblems again. I also made a change to remove a component from `updatesNeeded` before decomposing its arguments. This is to avoid concurrent modification (e.g. adding a subproblem while iterating through the mutable map of subproblems). Additionally, I made some more subtle changes to account for the possibility that a subproblem becomes unused because its corresponding parent value gets removed (for example, this could happen if refining with a binning strategy for continuous elements).